rutas api

+++ ProyectoBack/routes/api.js
@@ -1,7 +1,9 @@
 const router = require('express').Router();
 const Post = require('../models/Post');
+const moment = require('moment');
+const { check, validationResult } = require('express-validator');
 
-router.get('/', async (req, res) => {
+router.get('/posts', async (req, res) => {
     try {
         const posts = await Post.find();
         res.json(posts);
@@ -11,9 +13,39 @@ router.get('/', async (req, res) => {
     }
 });
 
-router.post('/', async (req, res) => {
+router.get('/posts/:id', check("id").isMongoId(), async (req, res) => {
+
+    const erroresValidacion = validationResult(req);
+    if (!erroresValidacion.isEmpty()) {
+        return res.json(erroresValidacion.array());
+    }
+
     try {
-        const post = await Post.create(req.body);
+        console.log(req.params.id);
+        const post = await Post.findById(req.params.id);
+        console.log(post);
+        res.json(post);
+    } catch (error) {
+        console.log("Error obtaining the post. ", error);
+        res.status(500).json({ error: error.message })
+    }
+});
+
+router.post('/posts', [
+    check('title').notEmpty(),
+    check('description').notEmpty(),
+    check('author').notEmpty(),
+], async (req, res) => {
+
+    const erroresValidacion = validationResult(req);
+    if (!erroresValidacion.isEmpty()) {
+        return res.json(erroresValidacion.array());
+    }
+
+    try {
+        let newPost = req.body;
+        newPost.date = moment().toDate();
+        const post = await Post.create(newPost);
         res.json({ success: "Post created", post: post });
     } catch (error) {
         console.log("Error creating the posts. ", error);
@@ -21,7 +53,17 @@ router.post('/', async (req, res) => {
     }
 });
 
-router.put('/', async (req, res) => {
+router.put('/posts', [
+    check('title').notEmpty(),
+    check('description').notEmpty(),
+    check('author').notEmpty(),
+], async (req, res) => {
+
+    const erroresValidacion = validationResult(req);
+    if (!erroresValidacion.isEmpty()) {
+        return res.json(erroresValidacion.array());
+    }
+
     try {
         const post = await Post.findByIdAndUpdate(req.body._id, req.body, { new: true });
         res.json({ success: "Post updated", post: post });
@@ -31,7 +73,13 @@ router.put('/', async (req, res) => {
     }
 });
 
-router.delete('/:id', (req, res) => {
+router.delete('/posts/:id', check('id').isMongoId(), async (req, res) => {
+
+    const erroresValidacion = validationResult(req);
+    if (!erroresValidacion.isEmpty()) {
+        return res.json(erroresValidacion.array());
+    }
+
     try {
         const post = await Post.findByIdAndDelete(req.params.id);
         res.json({ success: "Post deleted", post: post });
